Front End
Front end for a copy of the game Wordle from https://www.powerlanguage.co.uk/wordle/

This was a small project completed to build knowledge of React.

At the moment the user will need to input the 5 letter word to start the game, so it's best done with a friend.
There is no validation of what is a "word" and what isn't.

Eventually a back-end will be created to host a small database of game links and the associated word so it can be shared with friends.